+++ src/constant/make-latest-strategy.ts
@@ -1,0 +1,5 @@
+export const IF_NEW = "if-new";
+export const SEMVER = "semver";
+export const LEGACY = "legacy";
+export const ALWAYS = "always";
+export const NEVER = "never";