+++ src/type/make-latest-strategy.ts
@@ -1,0 +1,6 @@
+export type MakeLatestStrategy =
+  | "if-new"
+  | "semver"
+  | "legacy"
+  | "always"
+  | "never";